requirements.txt
google-adk>=1.25.0
google-genai>=1.64.0
python-dotenv>=1.2.0
